GDPR-41 Repackaging Data Compliance specific classes
This decision was based off a discussion that we may
want to move out the data compliance specific functions
in Elite2 API into its own separate service.
