+++ src/main/java/net/syscon/elite/api/resource/impl/OffenderResourceImpl.java
@@ -13,12 +13,11 @@ import net.syscon.elite.security.AuthenticationFacade;
 import net.syscon.elite.security.VerifyOffenderAccess;
 import net.syscon.elite.service.*;
 import org.apache.commons.lang3.StringUtils;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import uk.gov.justice.hmpps.nomis.datacompliance.service.OffenderDataComplianceService;
 
 import javax.validation.constraints.NotNull;
 import java.time.LocalDate;
