--- src/main/java/net/syscon/elite/events/dto/OffenderDeletionEvent.java
@@ -1,15 +1,0 @@
-package net.syscon.elite.events.dto;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-@AllArgsConstructor
-@NoArgsConstructor
-@Data
-public class OffenderDeletionEvent {
-    private String offenderIdDisplay;
-}
-
