--- src/main/java/net/syscon/elite/events/publishers/OffenderPendingDeletionEventPusher.java
@@ -1,5 +1,0 @@
-package net.syscon.elite.events.publishers;
-
-public interface OffenderPendingDeletionEventPusher {
-    void sendEvent(final String offenderIdDisplay);
-}
