+++ src/main/java/uk/gov/justice/hmpps/nomis/datacompliance/events/dto/OffenderDeletionEvent.java
@@ -1,0 +1,15 @@
+package uk.gov.justice.hmpps.nomis.datacompliance.events.dto;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+@AllArgsConstructor
+@NoArgsConstructor
+@Data
+public class OffenderDeletionEvent {
+    private String offenderIdDisplay;
+}
+
