+++ src/main/java/uk/gov/justice/hmpps/nomis/datacompliance/events/publishers/OffenderPendingDeletionEventPusher.java
@@ -1,0 +1,5 @@
+package uk.gov.justice.hmpps.nomis.datacompliance.events.publishers;
+
+public interface OffenderPendingDeletionEventPusher {
+    void sendEvent(final String offenderIdDisplay);
+}
